Re-organise application tests (#609)
